+++ docs/fr_FR/changelog.md
@@ -1,6 +1,11 @@
-# Changelog HEALTBOX
+# Changelog beta HEALTBOX
 
 
+
+# 19/02/2023
+
+- Ajout d'un CRON personnalisable.
+- Ajout d'une action CURSOR pour changer de profil.
 
 # 31/01/2023